chore: (re)move protos and config.Provider
