--- api/keto/acl/admin/v1alpha1/admin_service.proto
@@ -1,16 +1,0 @@
-syntax = "proto3";
-
-package keto.acl.admin.v1alpha1;
-
-option go_package = "github.com/ory/keto/api/keto/acl/admin/v1alpha1;admin";
-option csharp_namespace = "Ory.Keto.Acl.Admin.V1Alpha1";
-option java_multiple_files = true;
-option java_outer_classname = "AdminServiceProto";
-option java_package = "sh.ory.keto.acl.admin.v1alpha1";
-option php_namespace = "Ory\\Keto\\Acl\\Admin\\V1alpha1";
-
-// The admin service for administrative tasks in the ACL system.
-service AdminService {
-  // ...RPCs are coming here in (near) future...
-}
-
