--- api/keto/acl/node/v1alpha1/node_service.proto
@@ -1,28 +1,0 @@
-syntax = "proto3";
-
-package keto.acl.node.v1alpha1;
-
-option go_package = "github.com/ory/keto/api/keto/acl/node/v1alpha1;node";
-option csharp_namespace = "Ory.Keto.Acl.Node.V1Alpha1";
-option java_multiple_files = true;
-option java_outer_classname = "NodeServiceProto";
-option java_package = "sh.ory.keto.acl.node.v1alpha1";
-option php_namespace = "Ory\\Keto\\Acl\\Node\\V1alpha1";
-
-// The internal service for intercommunication
-// between ACL servers in a cluster.
-//
-// Implementing this service principally allows any program
-// in the network to participate as a node in an ACL cluster.
-service NodeService {
-  // ...RPCs for inter-node-communication are coming here...
-  //
-  // e.g.:
-  //  - the expand API to load off deep checks to other nodes
-  //    - we can attach meta to CheckResponses to gain greater knowledge of relationships
-  //      and measure statistics to integrate with Prometheus and Grafana
-  //  - schema node election
-  //  - replace groupcache's normal http server with protobuf grpc protocol!!!
-  //    - https://github.com/mailgun/groupcache/blob/8672aac324c41ba3909ad69e1a76b678894d378b/http.go#L160
-}
-
