--- api/keto/rbac/README.md
@@ -1,2 +1,0 @@
-# Maybe coming soon
-Built upon [ACL](../acl).
