+++ cmd/expand/root.go
@@ -8,9 +8,9 @@ import (
 	"github.com/ory/x/cmdx"
 	"github.com/spf13/cobra"
 
-	acl "github.com/ory/keto/api/keto/acl/v1alpha1"
 	"github.com/ory/keto/cmd/client"
 	"github.com/ory/keto/internal/expand"
+	acl "github.com/ory/keto/proto/ory/keto/acl/v1alpha1"
 )
 
 const FlagMaxDepth = "max-depth"
