+++ cmd/relationtuple/create.go
@@ -8,7 +8,7 @@ import (
 	"os"
 	"path/filepath"
 
-	acl "github.com/ory/keto/api/keto/acl/v1alpha1"
+	acl "github.com/ory/keto/proto/ory/keto/acl/v1alpha1"
 
 	"github.com/ory/keto/internal/relationtuple"
