+++ cmd/relationtuple/get.go
@@ -3,9 +3,10 @@ package relationtuple
 import (
 	"fmt"
 
+	acl "github.com/ory/keto/proto/ory/keto/acl/v1alpha1"
+
 	"github.com/ory/x/flagx"
 
-	acl "github.com/ory/keto/api/keto/acl/v1alpha1"
 	"github.com/ory/keto/internal/relationtuple"
 
 	"github.com/spf13/pflag"
