+++ internal/check/handler.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"net/http"
 
-	"google.golang.org/grpc"
+	acl "github.com/ory/keto/proto/ory/keto/acl/v1alpha1"
 
-	acl "github.com/ory/keto/api/keto/acl/v1alpha1"
+	"google.golang.org/grpc"
 
 	"github.com/ory/keto/internal/relationtuple"
