+++ internal/expand/handler.go
@@ -5,9 +5,9 @@ import (
 	"net/http"
 	"strconv"
 
-	"google.golang.org/grpc"
+	acl "github.com/ory/keto/proto/ory/keto/acl/v1alpha1"
 
-	acl "github.com/ory/keto/api/keto/acl/v1alpha1"
+	"google.golang.org/grpc"
 
 	"github.com/ory/keto/internal/relationtuple"
