+++ internal/expand/tree.go
@@ -5,9 +5,9 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/pkg/errors"
+	acl "github.com/ory/keto/proto/ory/keto/acl/v1alpha1"
 
-	acl "github.com/ory/keto/api/keto/acl/v1alpha1"
+	"github.com/pkg/errors"
 
 	"github.com/ory/keto/internal/relationtuple"
 )
