+++ internal/relationtuple/handler.go
@@ -3,7 +3,8 @@ package relationtuple
 import (
 	"google.golang.org/grpc"
 
-	acl "github.com/ory/keto/api/keto/acl/v1alpha1"
+	acl "github.com/ory/keto/proto/ory/keto/acl/v1alpha1"
+
 	"github.com/ory/keto/internal/x"
 )
